+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 * Support taking screenshots
+* Breaking change: Remove `device.disable_keyboard_helpers!`
+  Use `device.settings.disable_keyboard_helpers!` instead
 
 # 1.5.7